[VY] Fix throughput and other minor bugs
Fix throughput values (and tranmissions) in non-persistent simulation when
updating start of transmission times for nodes that sense the channel is
busy.

Fix potential error with finding next node to transmit its packet when
time to transmit is pushed back to beyond 2*T.

Fix error with resetting busy channel counter for non-persistent sim.